rule out already-uploaded dists earlier
If --neo is meant to mean "only the first time a given distname is
uploaded" -- and I propose that it is -- then we need to count each
upload for a distname before ruling dists out based on author.

Without this patch, --neo with --user finds the *user's* first upload
of a given dist.  This can certainly be useful, but I don't think it's
the intent of --neo.